+++ src/dual/mod.rs
@@ -1,10 +1,6 @@
 pub mod dual1;
-// pub mod dual1;
 pub mod dual2;
 pub mod dual_py;
 pub mod linalg;
 pub mod linalg_f64;
-
-
-
 pub mod linalg_py;
