+++ src/lib.rs
@@ -12,12 +12,15 @@ use dual::dual2::Dual2;
 use dual::linalg_py::{dsolve1_py, dsolve2_py, fdsolve1_py, fdsolve2_py};
 
 pub mod splines;
-use splines::PPSpline;
+use splines::spline_py::{PPSplineF64, PPSplineDual, PPSplineDual2};
 
 #[pymodule]
 fn rateslibrs(_py: Python, m: &PyModule) -> PyResult<()> {
     m.add_class::<Dual>()?;
     m.add_class::<Dual2>()?;
+    m.add_class::<PPSplineF64>()?;
+    m.add_class::<PPSplineDual>()?;
+    m.add_class::<PPSplineDual2>()?;
 
     // m.add_class::<PPSpline<f64>>()?;
     m.add_function(wrap_pyfunction!(dsolve1_py, m)?)?;
