+++ src/splines/spline_py.rs
@@ -1,0 +1,67 @@
+use crate::splines::spline_f64::PPSpline;
+use crate::dual::dual1::Dual;
+use crate::dual::dual2::Dual2;
+
+use pyo3::prelude::*;
+use pyo3::types::PyType;
+use pyo3::exceptions::{PyTypeError, PyValueError};
+
+macro_rules! create_interface {
+    ($name: ident, $type: ident) => {
+        #[pyclass]
+        pub struct $name {
+            inner: PPSpline<$type>,
+        }
+        #[pymethods]
+        impl $name {
+            #[new]
+            pub fn new(k: usize, t: Vec<f64>) -> Self {
+                Self {
+                    inner: PPSpline::new(k, t),
+                }
+            }
+
+            #[getter]
+            fn n(&self) -> PyResult<usize> {
+                Ok(*self.inner.n())
+            }
+
+            #[getter]
+            fn k(&self) -> PyResult<usize> {
+                Ok(*self.inner.k())
+            }
+
+            #[getter]
+            fn t(&self) -> PyResult<Vec<f64>> {
+                Ok(self.inner.t().clone())
+            }
+
+            #[getter]
+            fn c(&self) -> PyResult<Option<Vec<$type>>> {
+                match self.inner.c() {
+                    Some(val) => Ok(Some(val.clone().into_raw_vec())),
+                    None => Ok(None)
+                }
+            }
+
+            pub fn csolve(
+                &mut self,
+                tau: Vec<f64>,
+                y: Vec<$type>,
+                left_n: usize,
+                right_n: usize,
+                allow_lsq: bool
+            ) {
+                self.inner.csolve(&tau, &y, left_n, right_n, allow_lsq)
+            }
+
+            pub fn ppev_single(&self, x: f64) -> $type {
+                self.inner.ppev_single(&x)
+            }
+        }
+    };
+}
+
+create_interface!(PPSplineF64, f64);
+create_interface!(PPSplineDual, Dual);
+create_interface!(PPSplineDual2, Dual2);
